Update the web version of the program to v0.2.1.
In this new version I have put a header with the buttons that redirect to my portfolio, to the project's GitHub repository, so that users can donate to me and the change theme button.

Also add a button that allows you to copy the generated passwords to the users' clipboard.

Finally, place a footer so that the version of the program and the copyright can be seen.